+++ Microsoft.Azure.Cosmos/src/Resource/QueryResponses/CosmosFeedResponseUtil.cs
@@ -79,5 +79,17 @@ namespace Microsoft.Azure.Cosmos
             get => this.Data;
             set => this.Data = value;
         }
+
+        public Collection<T> Users
+        {
+            get => this.Data;
+            set => this.Data = value;
+        }
+
+        public Collection<T> Permissions
+        {
+            get => this.Data;
+            set => this.Data = value;
+        }
     }
 }