+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/ChangeFeedResultSetIteratorTests.cs
@@ -7,12 +7,16 @@ namespace Microsoft.Azure.Cosmos.Tests
     using System;
     using System.Collections.Generic;
     using System.IO;
+    using System.Linq;
     using System.Net;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos.Client.Core.Tests;
+    using Microsoft.Azure.Cosmos.Query;
+    using Microsoft.Azure.Cosmos.Query.Core;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
+    using Newtonsoft.Json;
 
     [TestClass]
     public class ChangeFeedResultSetIteratorTests
@@ -219,6 +223,40 @@ namespace Microsoft.Azure.Cosmos.Tests
                 It.IsAny<Action<RequestMessage>>(),
                 It.IsAny<Func<ResponseMessage, ResponseMessage>>(),
                 It.IsAny<CancellationToken>()), Times.Once);
+        }
+
+        [TestMethod]
+        public async Task GetChangeFeedTokensAsyncReturnsOnePerPartitionKeyRange()
+        {
+            // Setting mock to have 3 ranges, to generate 3 tokens
+            MultiRangeMockDocumentClient documentClient = new MultiRangeMockDocumentClient();
+            CosmosClient client = MockCosmosUtil.CreateMockCosmosClient();
+            Mock<CosmosClientContext> mockContext = new Mock<CosmosClientContext>();
+            mockContext.Setup(x => x.ClientOptions).Returns(MockCosmosUtil.GetDefaultConfiguration());
+            mockContext.Setup(x => x.DocumentClient).Returns(documentClient);
+            mockContext.Setup(x => x.CosmosSerializer).Returns(MockCosmosUtil.Serializer);
+            mockContext.Setup(x => x.Client).Returns(client);
+            mockContext.Setup(x => x.CreateLink(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>())).Returns(UriFactory.CreateDocumentCollectionUri("test", "test"));
+
+            DatabaseCore db = new DatabaseCore(mockContext.Object, "test");
+            ContainerCore container = new ContainerCore(mockContext.Object, db, "test");
+            IEnumerable<string> tokens = await container.GetChangeFeedTokensAsync();
+            Assert.AreEqual(3, tokens.Count());
+
+            Routing.PartitionKeyRangeCache pkRangeCache = await documentClient.GetPartitionKeyRangeCacheAsync();
+            foreach (string token in tokens)
+            {
+                // Validate that each token represents a StandByFeedContinuationToken with a single Range
+                List<CompositeContinuationToken> deserialized = JsonConvert.DeserializeObject<List<CompositeContinuationToken>>(token);
+                Assert.AreEqual(1, deserialized.Count);
+                CompositeContinuationToken compositeToken = deserialized[0];
+
+                IReadOnlyList<Documents.PartitionKeyRange> rangesForTheToken = await pkRangeCache.TryGetOverlappingRangesAsync("", compositeToken.Range);
+                // Token represents one range
+                Assert.AreEqual(1, rangesForTheToken.Count);
+                Assert.AreEqual(rangesForTheToken[0].MinInclusive, compositeToken.Range.Min);
+                Assert.AreEqual(rangesForTheToken[0].MaxExclusive, compositeToken.Range.Max);
+            }
         }
 
         private class MultiRangeMockDocumentClient : MockDocumentClient